+++ src/Bundle/Content/Model/StaticContent.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2014 Symfony CMF
+ * (c) 2011-2015 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,23 +12,18 @@
 namespace Symfony\Cmf\Bundle\ContentBundle\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
-
 use Knp\Menu\NodeInterface;
-
 use Sonata\BlockBundle\Model\BlockInterface;
 use Symfony\Cmf\Bundle\CoreBundle\Translatable\TranslatableInterface;
 use Symfony\Cmf\Component\Routing\RouteObjectInterface;
 use Symfony\Cmf\Component\Routing\RouteReferrersInterface;
-
 use Symfony\Cmf\Bundle\CoreBundle\PublishWorkflow\PublishTimePeriodInterface;
 use Symfony\Cmf\Bundle\CoreBundle\PublishWorkflow\PublishableInterface;
-
 use Symfony\Cmf\Bundle\MenuBundle\Model\MenuNode;
 use Symfony\Cmf\Bundle\MenuBundle\Model\MenuNodeReferrersInterface;
-use Symfony\Component\Routing\Route;
 
 /**
- * Standard implementation of StaticContent:
+ * Standard implementation of StaticContent:.
  *
  * Standard features:
  *
@@ -50,7 +45,7 @@ class StaticContent extends StaticContentBase implements
     TranslatableInterface
 {
     /**
-     * @var boolean whether this content is publishable
+     * @var bool whether this content is publishable
      */
     protected $publishable = true;
 
@@ -75,7 +70,7 @@ class StaticContent extends StaticContentBase implements
     protected $routes;
 
     /**
-     * MenuNode[]
+     * @var MenuNode[]
      */
     protected $menuNodes;
 
@@ -87,6 +82,8 @@ class StaticContent extends StaticContentBase implements
     /**
      * Hashmap for application data associated to this document. Both keys and
      * values must be strings.
+     *
+     * @var array
      */
     protected $extras = array();
 
@@ -105,7 +102,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function getLocale()
     {
@@ -113,7 +110,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function setLocale($locale)
     {
@@ -161,7 +158,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function setPublishable($publishable)
     {
@@ -169,7 +166,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function isPublishable()
     {
@@ -177,7 +174,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function getPublishStartDate()
     {
@@ -185,7 +182,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function setPublishStartDate(\DateTime $publishStartDate = null)
     {
@@ -193,7 +190,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function getPublishEndDate()
     {
@@ -201,7 +198,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function setPublishEndDate(\DateTime $publishEndDate = null)
     {
@@ -209,7 +206,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * Get the application information associated with this document
+     * Get the application information associated with this document.
      *
      * @return array
      */
@@ -219,7 +216,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * Get a single application information value
+     * Get a single application information value.
      *
      * @param string      $name
      * @param string|null $default
@@ -232,7 +229,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * Set the application information
+     * Set the application information.
      *
      * @param array $extras
      *
@@ -265,7 +262,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function addRoute($route)
     {
@@ -273,7 +270,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function removeRoute($route)
     {
@@ -281,7 +278,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function getRoutes()
     {
@@ -289,7 +286,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function addMenuNode(NodeInterface $menu)
     {
@@ -297,7 +294,7 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function removeMenuNode(NodeInterface $menu)
     {
@@ -305,11 +302,10 @@ class StaticContent extends StaticContentBase implements
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function getMenuNodes()
     {
         return $this->menuNodes;
     }
-
 }