+++ src/Bundle/Content/Tests/Functional/Doctrine/Phpcr/StaticContentTest.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2014 Symfony CMF
+ * (c) 2011-2015 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -34,7 +34,7 @@ class StaticContentTest extends BaseTestCase
             'publishEndDate' => new \DateTime('2013-06-18'),
         );
 
-        $content = new StaticContent;
+        $content = new StaticContent();
         $refl = new \ReflectionClass($content);
 
         $content->setParentDocument($this->base);