+++ src/Bundle/Content/Tests/Resources/DataFixtures/Phpcr/LoadContentData.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2014 Symfony CMF
+ * (c) 2011-2015 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -24,19 +24,19 @@ class LoadContentData implements FixtureInterface
         NodeHelper::createPath($manager->getPhpcrSession(), '/test');
         $root = $manager->find(null, '/test');
 
-        $contentRoot = new Generic;
+        $contentRoot = new Generic();
         $contentRoot->setNodename('contents');
         $contentRoot->setParent($root);
         $manager->persist($contentRoot);
 
-        $content = new StaticContent;
+        $content = new StaticContent();
         $content->setName('content-1');
         $content->setTitle('Content 1');
         $content->setBody('Content 1');
         $content->setParentDocument($contentRoot);
         $manager->persist($content);
 
-        $content = new StaticContent;
+        $content = new StaticContent();
         $content->setName('content-2');
         $content->setTitle('Content 2');
         $content->setBody('Content 2');